requirements.txt
pandas
pyspark
jupyter
azure-eventhub